+++ src/frontend/packages/lib_classroom/src/components/DashboardClassroomInstructor/index.tsx
@@ -1,7 +1,6 @@
 import { Button } from '@openfun/cunningham-react';
-import { Grid } from 'grommet';
 import { Breakpoints, Nullable } from 'lib-common';
-import { Box, BoxLoader, Classroom, useResponsive } from 'lib-components';
+import { Box, BoxLoader, Classroom, Grid, useResponsive } from 'lib-components';
 import React, { Suspense, lazy } from 'react';
 import { toast } from 'react-hot-toast';
 import { defineMessages, useIntl } from 'react-intl';